services/zoe-mcp-server/requirements.txt
# Zoe MCP Server Dependencies
mcp>=1.0.0
httpx>=0.25.0
PyJWT>=2.8.0
fastapi>=0.104.0
uvicorn>=0.24.0
